+++ backend/core/src/auth/controllers/user.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  Header,
   Param,
   ParseUUIDPipe,
   Post,
@@ -43,13 +44,19 @@ import { UserProfileAuthzGuard } from "../guards/user-profile-authz.guard"
 import { ActivityLogInterceptor } from "../../activity-log/interceptors/activity-log.interceptor"
 import { IdDto } from "../../shared/dto/id.dto"
 import { Response as ExpressResponse } from "express"
+import { UserCsvExporterService } from "../services/user-csv-exporter.service"
+import { Compare } from "../../shared/dto/filter.dto"
+
 @Controller("user")
 @ApiBearerAuth()
 @ApiTags("user")
 @ResourceType("user")
 @UsePipes(new ValidationPipe(defaultValidationPipeOptions))
 export class UserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly userCsvExporter: UserCsvExporterService
+  ) {}
 
   @Get()
   @UseGuards(DefaultAuthGuard, UserProfileAuthzGuard)
@@ -148,6 +155,24 @@ export class UserController {
   @ApiOperation({ summary: "List users", operationId: "list" })
   async list(@Query() queryParams: UserListQueryParams): Promise<PaginatedUserListDto> {
     return mapTo(PaginatedUserListDto, await this.userService.list(queryParams))
+  }
+
+  @Get("/csv")
+  @UseGuards(OptionalAuthGuard, AdminOrJurisdictionalAdminGuard)
+  @ApiOperation({ summary: "List users in CSV", operationId: "listAsCsv" })
+  @Header("Content-Type", "text/csv")
+  async listAsCsv(): Promise<string> {
+    const users = await this.userService.list({
+      page: 1,
+      limit: 300,
+      filter: [
+        {
+          isPortalUser: true,
+          $comparison: Compare["="],
+        },
+      ],
+    })
+    return this.userCsvExporter.exportFromObject(users)
   }
 
   @Post("/invite")