+++ backend/core/types/src/backend-swagger.ts
@@ -1103,6 +1103,21 @@ export class UserService {
     })
   }
   /**
+   * List users in CSV
+   */
+  listAsCsv(options: IRequestOptions = {}): Promise<string> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/user/csv"
+
+      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
+
+      let data = null
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
    * Invite user
    */
   invite(
@@ -1619,7 +1634,7 @@ export class MultiselectQuestionsService {
     })
   }
   /**
-   * Get multiselect question by id
+   * Get Listings by multiselect question id
    */
   retrieveListings(
     params: {