+++ sites/public/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.1](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@7.4.0...@bloom-housing/public@7.4.1) (2023-03-13)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+# [7.4.0](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@7.3.10...@bloom-housing/public@7.4.0) (2023-03-08)
+
+
+### Features
+
+* upgrade to node 18 and next 12 ([#3343](https://github.com/bloom-housing/bloom/issues/3343)) ([ee7bf77](https://github.com/bloom-housing/bloom/commit/ee7bf77dc3a46ee0795e3e6c4ca282960ec27394))
+
+
+
+
+
+## [7.3.10](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@7.3.9...@bloom-housing/public@7.3.10) (2023-03-08)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
 ## [7.3.9](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@7.3.8...@bloom-housing/public@7.3.9) (2023-02-28)
 
 **Note:** Version bump only for package @bloom-housing/public